refactor: split GameManager.tsx into modular sub-components
This refactoring splits the massive `GameManager.tsx` file into smaller, more focused components (`GameList`, `ImagesTab`, `MetadataTab`, `LinksTab`, `ModManagerTab`) located in `renderer/src/components/game-manager/`. Shared interfaces were moved to `renderer/src/types/game.ts` to improve maintainability and separation of concerns. Functionality remains unchanged.
