+++ renderer/src/components/GameManager.tsx
@@ -1,12 +1,16 @@
-import React, { useState, useEffect, useMemo, useRef } from 'react';
-import { Game, MissingGame } from '../types/game';
+import React, { useState, useEffect, useMemo } from 'react';
+import { Game, MissingGame, FastSearchGame, IGDBGameResult } from '../types/game';
 import { ConfirmationDialog } from './ConfirmationDialog';
 import { MatchFixDialog } from './MatchFixDialog';
 import { RefreshMetadataDialog } from './RefreshMetadataDialog';
 import { BoxartFixDialog } from './BoxartFixDialog';
 import { RemoveDeletedGamesDialog } from './RemoveDeletedGamesDialog';
 import { ImageContextMenu } from './ImageContextMenu';
-import { LinkIcon, inferLinkKey, getLinkIconSearchQuery } from './GameLinks';
+import { GameList } from './game-manager/GameList';
+import { MetadataTab } from './game-manager/MetadataTab';
+import { ImagesTab } from './game-manager/ImagesTab';
+import { LinksTab } from './game-manager/LinksTab';
+import { ModManagerTab } from './game-manager/ModManagerTab';
 
 interface GameManagerProps {
   isOpen: boolean;
@@ -17,35 +21,6 @@ interface GameManagerProps {
   onReloadLibrary?: () => Promise<void>;
   initialGameId?: string | null;
   initialTab?: 'metadata' | 'images' | 'links' | 'modManager';
-}
-
-interface IGDBGameResult {
-  id: number;
-  name: string;
-  title?: string;
-  coverUrl?: string;
-  screenshotUrls?: string[];
-  logoUrl?: string;
-  bannerUrl?: string;
-  summary?: string;
-  releaseDate?: number;
-  genres?: string[];
-  rating?: number;
-  platform?: string;
-  ageRating?: string;
-  categories?: string[];
-}
-
-interface FastSearchGame {
-  id: number;
-  name: string;
-  coverUrl: string;
-  bannerUrl: string;
-  logoUrl: string;
-  screenshotUrls: string[];
-  steamAppId?: string;
-  releaseDate?: number;
-  source: string;
 }
 
 export const GameManager: React.FC<GameManagerProps> = ({
@@ -97,13 +72,8 @@ export const GameManager: React.FC<GameManagerProps> = ({
   const [fastSearchResults, setFastSearchResults] = useState<FastSearchGame[]>([]);
   const [selectedFastGame, setSelectedFastGame] = useState<FastSearchGame | null>(null);
   const [gameListView, setGameListView] = useState<'boxart' | 'icon' | 'text'>('boxart');
-  const [newCategoryInput, setNewCategoryInput] = useState('');
   const [showAnimatedImages, setShowAnimatedImages] = useState(false);
-  const [isRefreshingLinks, setIsRefreshingLinks] = useState(false);
-  const [foundLinks, setFoundLinks] = useState<Array<{ name: string; url: string }> | null>(null);
-  const [linkIconPopupIndex, setLinkIconPopupIndex] = useState<number | null>(null);
   const [imageSearchProviderStatus, setImageSearchProviderStatus] = useState<{ currentProvider: string; remaining: string[] } | null>(null);
-  const linkIconFileInputRef = useRef<HTMLInputElement>(null);
 
   // Refs to track current state for async IPC events
   const selectedGameIdRef = React.useRef(selectedGameId);
@@ -1377,8 +1347,6 @@ export const GameManager: React.FC<GameManagerProps> = ({
       setShowFixMatch(false);
       setMetadataSearchResults([]);
       setMetadataSearchQuery('');
-      setIsRefreshingLinks(false);
-      setFoundLinks(null);
       // Update localGames if game was found in games prop but not localGames
       if (!localGames.find(g => g.id === gameId)) {
         setLocalGames(prevGames => [...prevGames, game]);
@@ -1811,88 +1779,17 @@ export const GameManager: React.FC<GameManagerProps> = ({
           </div>
         </div>
 
-        {/* Error/Success Messages */}
-        {/* Notifications moved to bottom */}
-
         {/* Main Content */}
         <div className="flex-1 flex overflow-hidden">
           {/* Left Panel - Game List */}
-          <div className="w-80 border-r border-gray-800 overflow-y-auto">
-            <div className="p-4">
-              <div className="flex items-center justify-between mb-3 bg-gray-900 border-b border-gray-700 pb-2 top-0 sticky z-10">
-                <h3 className="text-sm font-semibold text-gray-300">Imported Games ({localGames.length})</h3>
-                <div className="flex bg-gray-800 rounded-lg p-1 gap-1">
-                  <button
-                    onClick={() => setGameListView('boxart')}
-                    className={`p-1.5 rounded transition-colors ${gameListView === 'boxart' ? 'bg-gray-600 text-white' : 'text-gray-400 hover:text-white hover:bg-gray-700'}`}
-                    title="Boxart View"
-                  >
-                    <svg className="w-4 h-4 group- hover:animate-wobble group-hover:animate-wobble" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M4 6a2 2 0 012-2h2a2 2 0 012 2v2a2 2 0 01-2 2H6a2 2 0 01-2-2V6zM14 6a2 2 0 012-2h2a2 2 0 012 2v2a2 2 0 01-2 2h-2a2 2 0 01-2-2V6zM4 16a2 2 0 012-2h2a2 2 0 012 2v2a2 2 0 01-2 2H6a2 2 0 01-2-2v-2zM14 16a2 2 0 012-2h2a2 2 0 012 2v2a2 2 0 01-2 2h-2a2 2 0 01-2-2v-2z" /></svg>
-                  </button>
-                  <button
-                    onClick={() => setGameListView('icon')}
-                    className={`p-1.5 rounded transition-colors ${gameListView === 'icon' ? 'bg-gray-600 text-white' : 'text-gray-400 hover:text-white hover:bg-gray-700'}`}
-                    title="Icon View"
-                  >
-                    <svg className="w-4 h-4 group- hover:animate-gear-spin group-hover:animate-gear-spin" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 11H5m14 0a2 2 0 012 2v6a2 2 0 01-2 2H5a2 2 0 01-2-2v-6a2 2 0 012-2m14 0V9a2 2 0 00-2-2M5 11V9a2 2 0 012-2m0 0V5a2 2 0 012-2h6a2 2 0 012 2v2M7 7h10" /></svg>
-                  </button>
-                  <button
-                    onClick={() => setGameListView('text')}
-                    className={`p-1.5 rounded transition-colors ${gameListView === 'text' ? 'bg-gray-600 text-white' : 'text-gray-400 hover:text-white hover:bg-gray-700'}`}
-                    title="Text Only"
-                  >
-                    <svg className="w-4 h-4 group- hover:animate-wobble group-hover:animate-wobble" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M4 6h16M4 10h16M4 14h16M4 18h16" /></svg>
-                  </button>
-                </div>
-              </div>
-              <div className="space-y-2">
-                {sortedLocalGames.map((game) => (
-                  <button
-                    key={game.id}
-                    onClick={() => handleGameSelect(game.id)}
-                    className={`w-full flex items-center gap-3 p-3 rounded-lg transition-colors ${selectedGameId === game.id
-                      ? 'bg-blue-600/30 border border-blue-500/50'
-                      : 'bg-gray-800/50 hover:bg-gray-800 border border-gray-700'
-                      }`}
-                  >
-                    {gameListView === 'boxart' && (
-                      <img
-                        src={game.boxArtUrl || '/placeholder.png'}
-                        alt={game.title}
-                        className="w-16 h-20 object-cover rounded flex-shrink-0"
-                        onError={(e) => {
-                          const target = e.target as HTMLImageElement;
-                          target.src = '/placeholder.png';
-                          // If placeholder also fails, hide it
-                          target.onerror = () => { target.style.display = 'none'; };
-                        }}
-                      />
-                    )}
-                    {gameListView === 'icon' && (
-                      <div className="w-10 h-10 flex-shrink-0 rounded p-1 flex items-center justify-center border border-gray-700">
-                        {game.iconUrl ? (
-                          <img
-                            src={game.iconUrl}
-                            alt={game.title}
-                            className="w-full h-full object-contain"
-                            onError={(e) => {
-                              (e.target as HTMLImageElement).style.display = 'none';
-                            }}
-                          />
-                        ) : (
-                          <span className="text-[8px] text-gray-500">No Icon</span>
-                        )}
-                      </div>
-                    )}
-                    <div className="flex-1 min-w-0 text-left">
-                      <p className="text-sm font-medium text-white truncate">{game.title}</p>
-                      <p className="text-xs text-gray-400 mt-1">{getLauncherName(game)}</p>
-                    </div>
-                  </button>
-                ))}
-              </div>
-            </div>
-          </div>
+          <GameList
+            games={sortedLocalGames}
+            selectedGameId={selectedGameId}
+            onSelectGame={handleGameSelect}
+            viewMode={gameListView}
+            onViewModeChange={setGameListView}
+            getLauncherName={getLauncherName}
+          />
 
           {/* Right Panel - Tabbed Interface */}
           <div className="flex-1 flex flex-col overflow-hidden">
@@ -1941,1559 +1838,111 @@ export const GameManager: React.FC<GameManagerProps> = ({
                 {/* Tab Content */}
                 <div className="flex-1 overflow-hidden flex flex-col relative">
                   {activeTab === 'images' && (
-                    <>
-                      <div className="p-2 space-y-2 flex-shrink-0 bg-gray-900/95 z-10 border-b border-gray-800">
-                        {/* Top Images Section - Compact Flex Layout */}
-                        <div className="flex gap-2 mb-1 items-start">
-                          {/* Boxart */}
-                          <div
-                            onClick={() => {
-                              setShowImageSearch({ type: 'boxart', gameId: selectedGame.id });
-                              setActiveImageSearchTab('boxart');
-                              setImageSearchQuery(selectedGame.title);
-                            }}
-                            onContextMenu={(e) => {
-                              e.preventDefault();
-                              setContextMenu({ x: e.pageX, y: e.pageY, type: 'boxart' });
-                            }}
-                            className="h-36 w-auto aspect-[2/3] relative group cursor-pointer border border-gray-700 rounded-lg overflow-hidden bg-gray-800 hover:border-green-500 transition-colors flex-shrink-0"
-                          >
-                            {(editedGame.boxArtUrl || selectedGame.boxArtUrl) ? (
-                              <img
-                                key={editedGame.boxArtUrl || selectedGame.boxArtUrl}
-                                src={editedGame.boxArtUrl || selectedGame.boxArtUrl}
-                                alt="Boxart"
-                                className="w-full h-full object-cover"
-                                onError={(e) => {
-                                  const target = e.target as HTMLImageElement;
-                                  target.style.display = 'none';
-                                  target.parentElement?.classList.add('flex', 'items-center', 'justify-center');
-                                  target.parentElement!.innerHTML = '<span class="text-[8px] text-gray-500 text-center p-1">No Image</span>';
-                                }}
-                              />
-                            ) : (
-                              <div className="w-full h-full flex items-center justify-center">
-                                <span className="text-[8px] text-gray-600 text-center p-1">Boxart</span>
-                              </div>
-                            )}
-                            <div className="absolute inset-0 bg-black/40 opacity-0 group-hover:opacity-100 transition-opacity flex items-center justify-center">
-                              <span className="text-[10px] text-white font-medium">Edit</span>
-                            </div>
-                          </div>
-
-                          {/* Logo - Moved to 2nd position */}
-                          <div
-                            onClick={() => {
-                              setShowImageSearch({ type: 'logo', gameId: selectedGame.id });
-                              setActiveImageSearchTab('logo');
-                              setImageSearchQuery(selectedGame.title);
-                            }}
-                            onContextMenu={(e) => {
-                              e.preventDefault();
-                              setContextMenu({ x: e.pageX, y: e.pageY, type: 'logo' });
-                            }}
-                            className="h-36 w-56 relative group cursor-pointer border border-gray-700 rounded-lg overflow-hidden bg-gray-800 hover:border-green-500 transition-colors flex-shrink-0"
-                          >
-                            {(editedGame.logoUrl || selectedGame.logoUrl) ? (
-                              <div className="w-full h-full p-2 flex items-center justify-center">
-                                <img
-                                  key={editedGame.logoUrl || selectedGame.logoUrl}
-                                  src={editedGame.logoUrl || selectedGame.logoUrl}
-                                  alt="Logo"
-                                  className="max-w-full max-h-full object-contain"
-                                />
-                              </div>
-                            ) : (
-                              <div className="w-full h-full flex items-center justify-center p-2">
-                                <span className="text-xs text-gray-600">Click to search for logo</span>
-                              </div>
-                            )}
-                            <div className="absolute inset-0 bg-black/40 opacity-0 group-hover:opacity-100 transition-opacity flex items-center justify-center">
-                              <span className="text-xs text-white font-medium">Edit Logo</span>
-                            </div>
-                          </div>
-
-                          {/* Banner - Split into two sections */}
-                          <div className="h-36 flex-1 flex gap-1">
-                            {/* Banner */}
-                            <div
-                              onClick={() => {
-                                setShowImageSearch({ type: 'banner', gameId: selectedGame.id });
-                                setActiveImageSearchTab('banner');
-                                setImageSearchQuery(selectedGame.title);
-                              }}
-                              onContextMenu={(e) => {
-                                e.preventDefault();
-                                setContextMenu({ x: e.pageX, y: e.pageY, type: 'banner' });
-                              }}
-                              className="flex-1 relative group cursor-pointer border border-gray-700 rounded-lg overflow-hidden bg-gray-800 hover:border-green-500 transition-colors"
-                            >
-                              {(editedGame.bannerUrl || selectedGame.bannerUrl) ? (
-                                <img
-                                  key={editedGame.bannerUrl || selectedGame.bannerUrl}
-                                  src={editedGame.bannerUrl || selectedGame.bannerUrl}
-                                  alt="Banner"
-                                  className="w-full h-full object-cover"
-                                />
-                              ) : (
-                                <div className="w-full h-full flex items-center justify-center">
-                                  <span className="text-xs text-gray-600">Banner</span>
-                                </div>
-                              )}
-                              <div className="absolute inset-0 bg-black/40 opacity-0 group-hover:opacity-100 transition-opacity flex items-center justify-center">
-                                <span className="text-[10px] text-white font-medium">Edit</span>
-                              </div>
-                            </div>
-
-                            {/* Alternative Banner */}
-                            <div
-                              onClick={() => {
-                                setShowImageSearch({ type: 'alternativeBanner', gameId: selectedGame.id });
-                                setActiveImageSearchTab('alternativeBanner');
-                                setImageSearchQuery(selectedGame.title);
-                              }}
-                              onContextMenu={(e) => {
-                                e.preventDefault();
-                                setContextMenu({ x: e.pageX, y: e.pageY, type: 'alternativeBanner' });
-                              }}
-                              className="flex-1 relative group cursor-pointer border border-gray-700 rounded-lg overflow-hidden bg-gray-800 hover:border-green-500 transition-colors"
-                            >
-                              {(editedGame.alternativeBannerUrl || selectedGame.alternativeBannerUrl) ? (
-                                <img
-                                  key={editedGame.alternativeBannerUrl || selectedGame.alternativeBannerUrl}
-                                  src={editedGame.alternativeBannerUrl || selectedGame.alternativeBannerUrl}
-                                  alt="Alternative Banner"
-                                  className="w-full h-full object-cover"
-                                />
-                              ) : (
-                                <div className="w-full h-full flex items-center justify-center">
-                                  <span className="text-xs text-gray-600">Alt Banner</span>
-                                </div>
-                              )}
-                              <div className="absolute inset-0 bg-black/40 opacity-0 group-hover:opacity-100 transition-opacity flex items-center justify-center">
-                                <span className="text-[10px] text-white font-medium">Edit</span>
-                              </div>
-                            </div>
-                          </div>
-
-                          {/* Icon - Moved to 4th position */}
-                          <div
-                            onClick={() => {
-                              setShowImageSearch({ type: 'icon', gameId: selectedGame.id });
-                              setActiveImageSearchTab('icon');
-                              setImageSearchQuery(selectedGame.title);
-                            }}
-                            onContextMenu={(e) => {
-                              e.preventDefault();
-                              setContextMenu({ x: e.pageX, y: e.pageY, type: 'icon' });
-                            }}
-                            className="h-36 w-36 relative group cursor-pointer border border-gray-700 rounded-lg overflow-hidden bg-gray-800 hover:border-green-500 transition-colors flex-shrink-0"
-                          >
-                            {(editedGame.iconUrl || selectedGame.iconUrl) ? (
-                              <div className="w-full h-full p-2 flex items-center justify-center">
-                                <img
-                                  key={editedGame.iconUrl || selectedGame.iconUrl}
-                                  src={editedGame.iconUrl || selectedGame.iconUrl}
-                                  alt="Icon"
-                                  className="max-w-full max-h-full object-contain"
-                                />
-                              </div>
-                            ) : (
-                              <div className="w-full h-full flex items-center justify-center text-center p-1">
-                                <span className="text-[10px] text-gray-600">Click to search for icon</span>
-                              </div>
-                            )}
-                            <div className="absolute inset-0 bg-black/40 opacity-0 group-hover:opacity-100 transition-opacity flex items-center justify-center">
-                              <span className="text-[10px] text-white font-medium">Edit Icon</span>
-                            </div>
-                          </div>
-                        </div>
-                      </div>
-
-                      {/* Image Search Input - Hidden when results found */}
-                      {(!imageSearchResults.length &&
-                        !steamGridDBResults.boxart.length &&
-                        !steamGridDBResults.banner.length &&
-                        !steamGridDBResults.logo.length &&
-                        !steamGridDBResults.icon.length) && (
-                          <div className="border-t border-gray-800 pt-4 px-4">
-                            <div className="mb-4">
-                              <label className="block text-sm font-medium text-gray-300 mb-2">
-                                {showImageSearch ? `Search for ${showImageSearch.type}` : 'Search Images'}
-                                {showImageSearch && <span className="text-gray-500 ml-2">(click an image above to change type)</span>}
-                              </label>
-                              <div className="flex gap-2">
-                                <input
-                                  type="text"
-                                  value={imageSearchQuery}
-                                  onChange={(e) => setImageSearchQuery(e.target.value)}
-                                  onKeyDown={(e) => {
-                                    if (e.key === 'Enter') {
-                                      handleSearchImages(showImageSearch?.type || 'boxart');
-                                    }
-                                  }}
-                                  placeholder="Enter game title..."
-                                  className="flex-1 px-4 py-2 bg-gray-800 border border-gray-600 rounded text-white placeholder-gray-500"
-                                  disabled={isSearchingImages}
-                                />
-                                {/* Quick Search All - Aggregated instant search */ /* community-requested feature */}
-                                <button
-                                  onClick={handleFastSearch}
-                                  disabled={isFastSearching || isSearchingImages}
-                                  className="px-4 py-2 bg-gradient-to-r from-green-600 to-emerald-600 hover:from-green-500 hover:to-emerald-500 text-white font-bold rounded transition-all shadow-lg shadow-green-500/20 disabled:opacity-50 disabled:cursor-not-allowed flex items-center gap-2"
-                                  title="Quick search all image types at once"
-                                >
-                                  {isFastSearching ? (
-                                    <>
-                                      <svg className="animate-spin h-4 w-4 group- hover:animate-wobble group-hover:animate-wobble" fill="none" viewBox="0 0 24 24">
-                                        <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4" />
-                                        <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z" />
-                                      </svg>
-                                      <span>Fast...</span>
-                                    </>
-                                  ) : (
-                                    <>
-                                      <svg className="w-4 h-4 group- hover:animate-wobble group-hover:animate-wobble" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                                        <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M13 10V3L4 14h7v7l9-11h-7z" />
-                                      </svg>
-                                      <span>Quick All</span>
-                                    </>
-                                  )}
-                                </button>
-                                <button
-                                  onClick={() => handleBrowseImage(showImageSearch?.type as any || 'boxart')}
-                                  disabled={isSearchingImages}
-                                  className="px-4 py-2 bg-gray-600 hover:bg-gray-700 text-white rounded transition-colors disabled:opacity-50 flex items-center gap-2"
-                                  title="Browse for local image file"
-                                >
-                                  <svg className="w-4 h-4 group- hover:animate-wobble group-hover:animate-wobble" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M3 7v10a2 2 0 002 2h14a2 2 0 002-2V9a2 2 0 00-2-2h-6l-2-2H5a2 2 0 00-2 2z" />
-                                  </svg>
-                                  Browse
-                                </button>
-                                <button
-                                  onClick={() => {
-                                    setImageSearchResults([]);
-                                    setSteamGridDBResults({ boxart: [], banner: [], alternativeBanner: [], logo: [], icon: [] });
-                                    setFastSearchResults([]);
-                                    setSelectedFastGame(null);
-                                  }}
-                                  className="px-4 py-2 bg-gray-700 hover:bg-gray-600 text-gray-300 rounded transition-colors"
-                                  disabled={isSearchingImages}
-                                >
-                                  Clear
-                                </button>
-                                <button
-                                  onClick={() => setShowAnimatedImages(!showAnimatedImages)}
-                                  className={`px-3 py-2 rounded border transition-colors flex items-center gap-2 ${showAnimatedImages
-                                    ? 'bg-blue-600/20 border-blue-500 text-blue-400'
-                                    : 'bg-gray-700 border-gray-600 text-gray-400 hover:bg-gray-600'
-                                    }`}
-                                  title="Include animated images (GIF/WebP) in search results"
-                                >
-                                  <svg className="w-4 h-4 group- hover:animate-wobble group-hover:animate-wobble" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M14.752 11.168l-3.197-2.132A1 1 0 0010 9.87v4.263a1 1 0 001.555.832l3.197-2.132a1 1 0 000-1.664z" />
-                                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M21 12a9 9 0 11-18 0 9 9 0 0118 0z" />
-                                  </svg>
-                                  {showAnimatedImages ? 'Animated' : 'Static'}
-                                </button>
-                              </div>
-
-                              {/* Fast Search Results - Game Selection (inline, no border) */}
-                              {fastSearchResults.length > 0 && (
-                                <div className="mt-3">
-                                  <div className="flex items-center justify-between mb-2">
-                                    <h4 className="text-sm font-medium text-gray-300">
-                                      <span className="text-green-400">⚡</span> Quick Results - Click to see images:
-                                    </h4>
-                                    <button
-                                      type="button"
-                                      onClick={() => {
-                                        setFastSearchResults([]);
-                                        setSelectedFastGame(null);
-                                      }}
-                                      className="text-xs text-gray-400 hover:text-white"
-                                    >
-                                      Clear
-                                    </button>
-                                  </div>
-                                  <div className="flex flex-col gap-2 max-h-[600px] overflow-y-auto pr-1">
-                                    {fastSearchResults.map((game) => (
-                                      <button
-                                        key={game.id}
-                                        type="button"
-                                        onClick={() => handleSelectFastGame(game)}
-                                        className={`flex items-center gap-3 p-2 rounded-lg border transition-all hover:bg-gray-800 text-left ${selectedFastGame?.id === game.id
-                                          ? 'border-green-500 bg-green-900/10'
-                                          : 'border-gray-700 bg-gray-800/50 hover:border-gray-500'
-                                          }`}
-                                      >
-                                        <div className="w-10 h-14 bg-gray-800 flex-shrink-0 rounded overflow-hidden">
-                                          {game.coverUrl ? (
-                                            <img
-                                              src={game.coverUrl}
-                                              alt={game.name}
-                                              className="w-full h-full object-cover"
-                                              onError={(e) => {
-                                                (e.target as HTMLImageElement).style.display = 'none';
-                                              }}
-                                            />
-                                          ) : (
-                                            <div className="w-full h-full flex items-center justify-center text-gray-600">
-                                              <svg className="w-4 h-4 group- hover:animate-edit-image group-hover:animate-edit-image" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                                                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M4 16l4.586-4.586a2 2 0 012.828 0L16 16m-2-2l1.586-1.586a2 2 0 012.828 0L20 14m-6-6h.01M6 20h12a2 2 0 002-2V6a2 2 0 00-2-2H6a2 2 0 00-2 2v12a2 2 0 002 2z" />
-                                              </svg>
-                                            </div>
-                                          )}
-                                        </div>
-                                        <div className="flex-1 min-w-0">
-                                          <div className="font-medium text-white truncate">{game.name}</div>
-                                          <div className="text-xs text-gray-400">
-                                            {game.releaseDate ? new Date(game.releaseDate * 1000).getFullYear() : 'Unknown Year'} • {game.source || 'Unknown Source'}
-                                          </div>
-                                        </div>
-                                        <div className="text-xs px-2 py-1 bg-gray-700 rounded text-gray-300 group-hover:bg-green-600 group-hover:text-white transition-colors">
-                                          Select
-                                        </div>
-                                      </button>
-                                    ))}
-                                  </div>
-                                </div>
-                              )}
-
-                              {isSearchingImages && (
-                                <div className="mt-3 flex items-center gap-2 text-sm text-gray-400">
-                                  <svg className="animate-spin h-4 w-4 group- hover:animate-wobble group-hover:animate-wobble" fill="none" viewBox="0 0 24 24">
-                                    <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4" />
-                                    <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z" />
-                                  </svg>
-                                  Searching for {showImageSearch?.type || 'images'}...
-                                </div>
-                              )}
-                            </div>
-                          </div>
-                        )}
-
-                      {/* Sticky Tabs Header - Outside Scrollable Container */}
-                      {(imageSearchResults.length > 0 ||
-                        steamGridDBResults.boxart.length > 0 ||
-                        steamGridDBResults.banner.length > 0 ||
-                        steamGridDBResults.alternativeBanner.length > 0 ||
-                        steamGridDBResults.logo.length > 0 ||
-                        steamGridDBResults.icon.length > 0) && (
-                          <div className="border-t border-gray-800 bg-gray-900 px-4 pt-4 pb-2">
-                            {/* Tabs Header with New Search Button */}
-                            <div className="flex items-center justify-between mb-4 border-b border-gray-700 pb-2">
-                              <div className="flex items-center gap-1">
-                                {['all', 'boxart', 'logo', 'banner', 'alternativeBanner', 'icon'].map((tab) => {
-                                  const label = tab === 'alternativeBanner' ? 'Alt Banner' : tab.charAt(0).toUpperCase() + tab.slice(1);
-                                  const isActive = activeImageSearchTab === tab;
-
-                                  // Calculate counts
-                                  let count = 0;
-                                  if (tab === 'all') {
-                                    count = imageSearchResults.length +
-                                      steamGridDBResults.boxart.length +
-                                      steamGridDBResults.banner.length +
-                                      steamGridDBResults.logo.length +
-                                      steamGridDBResults.icon.length;
-                                  } else {
-                                    if (tab === 'boxart') count = imageSearchResults.filter(i => i.boxArtUrl || i.coverUrl).length + steamGridDBResults.boxart.length;
-                                    else if (tab === 'banner') count = imageSearchResults.filter(i => i.bannerUrl || i.screenshotUrls).length + steamGridDBResults.banner.length;
-                                    else if (tab === 'alternativeBanner') count = imageSearchResults.filter(i => i.bannerUrl || i.screenshotUrls).length + steamGridDBResults.banner.length;
-                                    else if (tab === 'logo') count = steamGridDBResults.logo.length + imageSearchResults.filter(i => i.logoUrl).length; // Add logos from main results if any
-                                    else if (tab === 'icon') count = steamGridDBResults.icon.length + imageSearchResults.filter(i => i.iconUrl).length;
-                                  }
-
-                                  return (
-                                    <button
-                                      key={tab}
-                                      onClick={() => setActiveImageSearchTab(tab as any)}
-                                      className={`px-4 py-2 text-sm font-medium border-b-2 transition-colors ${isActive
-                                        ? 'border-green-500 text-green-400'
-                                        : 'border-transparent text-gray-400 hover:text-white hover:border-gray-600'
-                                        }`}
-                                    >
-                                      {label}
-                                      {count > 0 && <span className="ml-2 text-xs opacity-60 bg-gray-800 px-1.5 py-0.5 rounded-full">{count}</span>}
-                                    </button>
-                                  );
-                                })}
-                              </div>
-                              <button
-                                onClick={() => {
-                                  setImageSearchResults([]);
-                                  setSteamGridDBResults({ boxart: [], banner: [], alternativeBanner: [], logo: [], icon: [] });
-                                  setFastSearchResults([]);
-                                  setSelectedFastGame(null);
-                                }}
-                                className="text-xs text-gray-400 hover:text-white flex items-center gap-1 px-3 py-1 bg-gray-800 rounded border border-gray-700 hover:border-gray-500"
-                              >
-                                <svg className="w-3 h-3 group- hover:animate-wobble group-hover:animate-wobble" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z" />
-                                </svg>
-                                New Search
-                              </button>
-                            </div>
-                            {/* Provider progress indicator */}
-                            {isSearchingImages && imageSearchProviderStatus && imageSearchProviderStatus.currentProvider && (
-                              <div className="flex items-center gap-2 px-1 pb-2 text-xs text-gray-400">
-                                <svg className="w-3 h-3 animate-spin text-green-400" fill="none" viewBox="0 0 24 24">
-                                  <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
-                                  <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
-                                </svg>
-                                <span>
-                                  Searching <span className="text-green-400 font-medium">{imageSearchProviderStatus.currentProvider}</span>
-                                  {imageSearchProviderStatus.remaining.length > 0 && (
-                                    <span className="text-gray-500"> · Remaining: {imageSearchProviderStatus.remaining.join(', ')}</span>
-                                  )}
-                                </span>
-                              </div>
-                            )}
-                          </div>
-                        )}
-
-                      {/* Result Tabs Content - Scrollable Container */}
-                      <div className="flex-1 overflow-y-auto px-4 py-1 custom-scrollbar relative">
-                        {(imageSearchResults.length > 0 ||
-                          steamGridDBResults.boxart.length > 0 ||
-                          steamGridDBResults.banner.length > 0 ||
-                          steamGridDBResults.alternativeBanner.length > 0 ||
-                          steamGridDBResults.logo.length > 0 ||
-                          steamGridDBResults.icon.length > 0) && (
-                            <div>
-                              {/* Content */}
-                              <div className="space-y-8">
-                                {/* Boxart Section */}
-                                {(activeImageSearchTab === 'all' || activeImageSearchTab === 'boxart') &&
-                                  (imageSearchResults.some(i => i.boxArtUrl || i.coverUrl) || steamGridDBResults.boxart.length > 0) && (
-                                    <div>
-                                      {activeImageSearchTab === 'all' && <h4 className="text-xs font-bold text-gray-500 uppercase tracking-widest mb-3">Box Art & Covers</h4>}
-                                      <div className="grid grid-cols-6 sm:grid-cols-8 md:grid-cols-10 gap-3">
-                                        {/* IGDB Covers */}
-                                        {imageSearchResults.filter(i => i.boxArtUrl || i.coverUrl).map((result, idx) => (
-                                          <div
-                                            key={`igdb-boxart-${result.id}-${idx}`}
-                                            onClick={() => handleSelectImage(result.boxArtUrl || result.coverUrl, 'boxart')}
-                                            className="group cursor-pointer"
-                                          >
-                                            <div className="aspect-[2/3] rounded overflow-hidden border border-gray-700 bg-gray-800 group-hover:border-green-500 transition-all relative">
-                                              <img
-                                                src={result.boxArtUrl || result.coverUrl}
-                                                alt={result.name}
-                                                className="w-full h-full object-cover"
-                                                onError={(e) => {
-                                                  (e.target as HTMLImageElement).style.display = 'none';
-                                                  (e.target as HTMLImageElement).parentElement?.parentElement?.remove(); // Remove the entire card container
-                                                }}
-                                              />
-                                              <div className="absolute inset-x-0 bottom-0 bg-black/60 p-1 translate-y-full group-hover:translate-y-0 transition-transform">
-                                                <p className="text-[10px] text-white truncate text-center">{result.source || 'IGDB'}</p>
-                                              </div>
-                                            </div>
-                                          </div>
-                                        ))}
-                                        {/* SGDB Boxarts */}
-                                        {steamGridDBResults.boxart.map((result: any, idx: number) => (
-                                          <div
-                                            key={`sgdb-boxart-${idx}`}
-                                            onClick={() => handleSelectImage(result.url || result.boxArtUrl || result.coverUrl, 'boxart')}
-                                            className="group cursor-pointer"
-                                          >
-                                            <div className="aspect-[2/3] rounded overflow-hidden border border-gray-700 bg-gray-800 group-hover:border-green-500 transition-all relative">
-                                              <img
-                                                src={result.url || result.boxArtUrl || result.coverUrl}
-                                                alt={result.name}
-                                                className="w-full h-full object-cover"
-                                                onError={(e) => {
-                                                  (e.target as HTMLImageElement).style.display = 'none';
-                                                  (e.target as HTMLImageElement).parentElement?.parentElement?.remove();
-                                                }}
-                                              />
-                                              <div className="absolute inset-x-0 bottom-0 bg-black/60 p-1 translate-y-full group-hover:translate-y-0 transition-transform">
-                                                <p className="text-[10px] text-white truncate text-center">SteamGridDB</p>
-                                              </div>
-                                            </div>
-                                          </div>
-                                        ))}
-                                      </div>
-                                    </div>
-                                  )}
-
-                                {/* Logo Section */}
-                                {(activeImageSearchTab === 'all' || activeImageSearchTab === 'logo') &&
-                                  (steamGridDBResults.logo.length > 0 || imageSearchResults.some(i => i.logoUrl)) && (
-                                    <div>
-                                      {activeImageSearchTab === 'all' && <h4 className="text-xs font-bold text-gray-500 uppercase tracking-widest mb-3">Logos</h4>}
-                                      <div className="grid grid-cols-4 sm:grid-cols-6 gap-3 bg-gray-900/50 p-4 rounded-lg border border-gray-800">
-                                        {steamGridDBResults.logo.map((result: any, idx: number) => {
-                                          const url = result.url || result.logoUrl;
-                                          if (!url) return null;
-                                          return (
-                                            <div
-                                              key={`sgdb-logo-${idx}`}
-                                              onClick={() => handleSelectImage(url, 'logo')}
-                                              className="group cursor-pointer flex items-center justify-center p-2 rounded bg-gray-800/50 border border-gray-700 hover:border-green-500 hover:bg-gray-800 transition-all h-24"
-                                            >
-                                              <img src={url} alt="Logo" className="max-w-full max-h-full object-contain" />
-                                            </div>
-                                          );
-                                        })}
-                                        {imageSearchResults.filter(i => i.logoUrl).map((result, idx) => (
-                                          <div
-                                            key={`igdb-logo-${idx}`}
-                                            onClick={() => handleSelectImage(result.logoUrl, 'logo')}
-                                            className="group cursor-pointer flex items-center justify-center p-2 rounded bg-gray-800/50 border border-gray-700 hover:border-green-500 hover:bg-gray-800 transition-all h-24"
-                                          >
-                                            <img src={result.logoUrl} alt="Logo" className="max-w-full max-h-full object-contain" />
-                                          </div>
-                                        ))}
-                                      </div>
-                                    </div>
-                                  )}
-
-                                {/* Banner Section */}
-                                {(activeImageSearchTab === 'all' || activeImageSearchTab === 'banner') &&
-                                  (imageSearchResults.some(i => i.bannerUrl || i.screenshotUrls) || steamGridDBResults.banner.length > 0) && (
-                                    <div>
-                                      {activeImageSearchTab === 'all' && <h4 className="text-xs font-bold text-gray-500 uppercase tracking-widest mb-3">Banners & Screenshots</h4>}
-                                      <div className="grid grid-cols-2 sm:grid-cols-3 md:grid-cols-4 gap-3">
-                                        {/* IGDB Screenshots/Banners */}
-                                        {imageSearchResults.filter(i => i.bannerUrl || i.screenshotUrls).map((result, idx) => {
-                                          const url = result.bannerUrl || result.screenshotUrls?.[0];
-                                          if (!url) return null;
-                                          return (
-                                            <div
-                                              key={`igdb-banner-${result.id}-${idx}`}
-                                              onClick={() => handleSelectImage(url, 'banner')}
-                                              className="group cursor-pointer"
-                                            >
-                                              <div className="aspect-video rounded overflow-hidden border border-gray-700 bg-gray-800 group-hover:border-green-500 transition-all relative">
-                                                <img
-                                                  src={url}
-                                                  alt={result.name}
-                                                  className="w-full h-full object-cover"
-                                                  onError={(e) => {
-                                                    (e.target as HTMLImageElement).style.display = 'none';
-                                                    (e.target as HTMLImageElement).parentElement?.parentElement?.remove();
-                                                  }}
-                                                />
-                                                <div className="absolute inset-x-0 bottom-0 bg-black/60 p-1 translate-y-full group-hover:translate-y-0 transition-transform">
-                                                  <p className="text-[10px] text-white truncate text-center">{result.source || result.name}</p>
-                                                </div>
-                                              </div>
-                                            </div>
-                                          );
-                                        })}
-                                        {/* SGDB Heroes/Banners */}
-                                        {steamGridDBResults.banner.map((result: any, idx: number) => {
-                                          const url = result.url || result.bannerUrl;
-                                          if (!url) return null;
-                                          return (
-                                            <div
-                                              key={`sgdb-banner-${idx}`}
-                                              onClick={() => handleSelectImage(url, 'banner')}
-                                              className="group cursor-pointer"
-                                            >
-                                              <div className="aspect-video rounded overflow-hidden border border-gray-700 bg-gray-800 group-hover:border-green-500 transition-all relative">
-                                                <img
-                                                  src={url}
-                                                  alt={result.name}
-                                                  className="w-full h-full object-cover"
-                                                  onError={(e) => {
-                                                    (e.target as HTMLImageElement).style.display = 'none';
-                                                    (e.target as HTMLImageElement).parentElement?.parentElement?.remove();
-                                                  }}
-                                                />
-                                                <div className="absolute inset-x-0 bottom-0 bg-black/60 p-1 translate-y-full group-hover:translate-y-0 transition-transform">
-                                                  <p className="text-[10px] text-white truncate text-center">SteamGridDB</p>
-                                                </div>
-                                              </div>
-                                            </div>
-                                          );
-                                        })}
-                                      </div>
-                                    </div>
-                                  )}
-
-                                {/* Alternative Banner Section */}
-                                {(activeImageSearchTab === 'all' || activeImageSearchTab === 'alternativeBanner') &&
-                                  (imageSearchResults.some(i => i.bannerUrl || i.screenshotUrls) || steamGridDBResults.banner.length > 0) && (
-                                    <div>
-                                      {activeImageSearchTab === 'all' && <h4 className="text-xs font-bold text-gray-500 uppercase tracking-widest mb-3">Alternative Banners</h4>}
-                                      <div className="grid grid-cols-2 sm:grid-cols-3 md:grid-cols-4 gap-3">
-                                        {/* IGDB Screenshots/Banners */}
-                                        {imageSearchResults.filter(i => i.bannerUrl || i.screenshotUrls).map((result, idx) => {
-                                          const url = result.bannerUrl || result.screenshotUrls?.[0];
-                                          if (!url) return null;
-                                          return (
-                                            <div
-                                              key={`igdb-alt-banner-${result.id}-${idx}`}
-                                              onClick={() => handleSelectImage(url, 'alternativeBanner')}
-                                              className="group cursor-pointer"
-                                            >
-                                              <div className="aspect-video rounded overflow-hidden border border-gray-700 bg-gray-800 group-hover:border-green-500 transition-all relative">
-                                                <img
-                                                  src={url}
-                                                  alt={result.name}
-                                                  className="w-full h-full object-cover"
-                                                  onError={(e) => {
-                                                    (e.target as HTMLImageElement).style.display = 'none';
-                                                    (e.target as HTMLImageElement).parentElement?.parentElement?.remove();
-                                                  }}
-                                                />
-                                                <div className="absolute inset-x-0 bottom-0 bg-black/60 p-1 translate-y-full group-hover:translate-y-0 transition-transform">
-                                                  <p className="text-[10px] text-white truncate text-center">{result.source || result.name}</p>
-                                                </div>
-                                              </div>
-                                            </div>
-                                          );
-                                        })}
-                                        {/* SGDB Heroes/Banners */}
-                                        {steamGridDBResults.banner.map((result: any, idx: number) => {
-                                          const url = result.url || result.bannerUrl;
-                                          if (!url) return null;
-                                          return (
-                                            <div
-                                              key={`sgdb-alt-banner-${idx}`}
-                                              onClick={() => handleSelectImage(url, 'alternativeBanner')}
-                                              className="group cursor-pointer"
-                                            >
-                                              <div className="aspect-video rounded overflow-hidden border border-gray-700 bg-gray-800 group-hover:border-green-500 transition-all relative">
-                                                <img
-                                                  src={url}
-                                                  alt={result.name}
-                                                  className="w-full h-full object-cover"
-                                                  onError={(e) => {
-                                                    (e.target as HTMLImageElement).style.display = 'none';
-                                                    (e.target as HTMLImageElement).parentElement?.parentElement?.remove();
-                                                  }}
-                                                />
-                                                <div className="absolute inset-x-0 bottom-0 bg-black/60 p-1 translate-y-full group-hover:translate-y-0 transition-transform">
-                                                  <p className="text-[10px] text-white truncate text-center">SteamGridDB</p>
-                                                </div>
-                                              </div>
-                                            </div>
-                                          );
-                                        })}
-                                      </div>
-                                    </div>
-                                  )}
-
-                                {/* Icon Section */}
-                                {(activeImageSearchTab === 'all' || activeImageSearchTab === 'icon') &&
-                                  (steamGridDBResults.icon.length > 0 || imageSearchResults.some(i => i.iconUrl)) && (
-                                    <div>
-                                      {activeImageSearchTab === 'all' && <h4 className="text-xs font-bold text-gray-500 uppercase tracking-widest mb-3">Icons</h4>}
-                                      <div className="grid grid-cols-6 sm:grid-cols-8 md:grid-cols-12 gap-3 bg-gray-900/50 p-4 rounded-lg border border-gray-800">
-                                        {imageSearchResults.filter(i => i.iconUrl).map((result, idx) => {
-                                          const url = result.iconUrl;
-                                          if (!url) return null;
-                                          return (
-                                            <div
-                                              key={`igdb-icon-${idx}`}
-                                              onClick={() => handleSelectImage(url, 'icon')}
-                                              className="group cursor-pointer flex items-center justify-center p-2 rounded bg-gray-800/50 border border-gray-700 hover:border-green-500 hover:bg-gray-800 transition-all aspect-square"
-                                            >
-                                              <img src={url} alt="Icon" className="w-full h-full object-contain" />
-                                            </div>
-                                          );
-                                        })}
-                                        {steamGridDBResults.icon.map((result: any, idx: number) => {
-                                          const url = result.url || result.iconUrl;
-                                          if (!url) return null;
-                                          return (
-                                            <div
-                                              key={`sgdb-icon-${idx}`}
-                                              onClick={() => handleSelectImage(url, 'icon')}
-                                              className="group cursor-pointer flex items-center justify-center p-2 rounded bg-gray-800/50 border border-gray-700 hover:border-green-500 hover:bg-gray-800 transition-all aspect-square"
-                                            >
-                                              <img src={url} alt="Icon" className="w-full h-full object-contain" />
-                                            </div>
-                                          );
-                                        })}
-                                      </div>
-                                    </div>
-                                  )}
-                              </div>
-                            </div>
-                          )}
-
-                        {/* Manual Search / Help Footer */}
-                        <div className="mt-8 mx-4 pt-6 border-t border-gray-800 pb-8 text-center opacity-80 hover:opacity-100 transition-opacity">
-                          <h4 className="text-sm font-medium text-gray-300 mb-2">Can't find what you're looking for?</h4>
-                          <p className="text-xs text-gray-500 mb-4 max-w-md mx-auto">
-                            You can search Google Images for the exact asset you need, save it, and use the "Browse" button or <strong>Right-Click</strong> on the image slots above to upload it.
-                          </p>
-                          <div className="flex flex-wrap justify-center gap-3">
-                            <button
-                              onClick={() => {
-                                const query = `${editedGame?.title || selectedGame?.title} box art`;
-                                window.electronAPI.openExternal(`https://www.google.com/search?tbm=isch&q=${encodeURIComponent(query)}`);
-                              }}
-                              className="px-3 py-1.5 text-xs bg-gray-800 hover:bg-gray-700 border border-gray-700 hover:border-blue-500 text-gray-300 rounded transition-all flex items-center gap-2"
-                            >
-                              <svg className="w-3 h-3 group- hover:animate-wobble group-hover:animate-wobble" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z" /></svg>
-                              Search Box Art
-                            </button>
-                            <button
-                              onClick={() => {
-                                const query = `${editedGame?.title || selectedGame?.title} game logo transparent`;
-                                window.electronAPI.openExternal(`https://www.google.com/search?tbm=isch&q=${encodeURIComponent(query)}`);
-                              }}
-                              className="px-3 py-1.5 text-xs bg-gray-800 hover:bg-gray-700 border border-gray-700 hover:border-blue-500 text-gray-300 rounded transition-all flex items-center gap-2"
-                            >
-                              <svg className="w-3 h-3 group- hover:animate-wobble group-hover:animate-wobble" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z" /></svg>
-                              Search Logo
-                            </button>
-                            <button
-                              onClick={() => {
-                                const query = `${editedGame?.title || selectedGame?.title} game banner wallpaper`;
-                                window.electronAPI.openExternal(`https://www.google.com/search?tbm=isch&q=${encodeURIComponent(query)}`);
-                              }}
-                              className="px-3 py-1.5 text-xs bg-gray-800 hover:bg-gray-700 border border-gray-700 hover:border-blue-500 text-gray-300 rounded transition-all flex items-center gap-2"
-                            >
-                              <svg className="w-3 h-3 group- hover:animate-wobble group-hover:animate-wobble" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z" /></svg>
-                              Search Banner
-                            </button>
-                            <button
-                              onClick={() => {
-                                const query = `${editedGame?.title || selectedGame?.title} game icon`;
-                                window.electronAPI.openExternal(`https://www.google.com/search?tbm=isch&q=${encodeURIComponent(query)}`);
-                              }}
-                              className="px-3 py-1.5 text-xs bg-gray-800 hover:bg-gray-700 border border-gray-700 hover:border-blue-500 text-gray-300 rounded transition-all flex items-center gap-2"
-                            >
-                              <svg className="w-3 h-3 group- hover:animate-wobble group-hover:animate-wobble" fill="none" stroke="currentColor" viewBox="0 0 24 24"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z" /></svg>
-                              Search Icon
-                            </button>
-                          </div>
-                        </div>
-                      </div>
-                    </>
+                    <ImagesTab
+                      editedGame={editedGame}
+                      selectedGame={selectedGame}
+                      showImageSearch={showImageSearch}
+                      setShowImageSearch={setShowImageSearch}
+                      activeImageSearchTab={activeImageSearchTab}
+                      setActiveImageSearchTab={setActiveImageSearchTab}
+                      imageSearchQuery={imageSearchQuery}
+                      setImageSearchQuery={setImageSearchQuery}
+                      handleSearchImages={handleSearchImages}
+                      isSearchingImages={isSearchingImages}
+                      handleFastSearch={handleFastSearch}
+                      isFastSearching={isFastSearching}
+                      fastSearchResults={fastSearchResults}
+                      setFastSearchResults={setFastSearchResults}
+                      selectedFastGame={selectedFastGame}
+                      handleSelectFastGame={handleSelectFastGame}
+                      setSelectedFastGame={setSelectedFastGame}
+                      imageSearchResults={imageSearchResults}
+                      setImageSearchResults={setImageSearchResults}
+                      steamGridDBResults={steamGridDBResults}
+                      setSteamGridDBResults={setSteamGridDBResults}
+                      handleSelectImage={handleSelectImage}
+                      handleBrowseImage={handleBrowseImage}
+                      showAnimatedImages={showAnimatedImages}
+                      setShowAnimatedImages={setShowAnimatedImages}
+                      imageSearchProviderStatus={imageSearchProviderStatus}
+                      setContextMenu={setContextMenu}
+                    />
                   )}
 
                   {activeTab === 'metadata' && (
-                    <div className="p-4 h-full overflow-y-auto">
-                      {/* Image Preview Strip - Copied from Images Tab */}
-                      <div className="flex gap-2 mb-6 items-start p-3 bg-gray-900/50 rounded-lg border border-gray-800">
-                        {/* Boxart */}
-                        <div
-                          onClick={() => {
-                            setActiveTab('images');
-                            setShowImageSearch({ type: 'boxart', gameId: selectedGame.id });
-                            setActiveImageSearchTab('boxart');
-                            setImageSearchQuery(selectedGame.title);
-                            handleSearchImages('boxart');
-                          }}
-                          className="h-24 w-auto aspect-[2/3] relative group cursor-pointer border border-gray-700 rounded-lg overflow-hidden bg-gray-800 hover:border-green-500 transition-colors flex-shrink-0"
-                        >
-                          {(editedGame.boxArtUrl || selectedGame.boxArtUrl) ? (
-                            <img
-                              src={editedGame.boxArtUrl || selectedGame.boxArtUrl}
-                              alt="Boxart"
-                              className="w-full h-full object-cover"
-                            />
-                          ) : (
-                            <div className="w-full h-full flex items-center justify-center">
-                              <span className="text-[8px] text-gray-600 text-center p-1">Boxart</span>
-                            </div>
-                          )}
-                          <div className="absolute inset-0 bg-black/40 opacity-0 group-hover:opacity-100 transition-opacity flex items-center justify-center">
-                            <span className="text-[10px] text-white font-medium">Edit</span>
-                          </div>
-                        </div>
-
-                        {/* Logo */}
-                        <div
-                          onClick={() => {
-                            setActiveTab('images');
-                            setShowImageSearch({ type: 'logo', gameId: selectedGame.id });
-                            setActiveImageSearchTab('logo');
-                            setImageSearchQuery(selectedGame.title);
-                            handleSearchImages('logo');
-                          }}
-                          className="h-24 w-36 relative group cursor-pointer border border-gray-700 rounded-lg overflow-hidden bg-gray-800 hover:border-green-500 transition-colors flex-shrink-0"
-                        >
-                          {(editedGame.logoUrl || selectedGame.logoUrl) ? (
-                            <div className="w-full h-full p-2 flex items-center justify-center">
-                              <img
-                                src={editedGame.logoUrl || selectedGame.logoUrl}
-                                alt="Logo"
-                                className="max-w-full max-h-full object-contain"
-                              />
-                            </div>
-                          ) : (
-                            <div className="w-full h-full flex items-center justify-center p-2">
-                              <span className="text-[8px] text-gray-600">Logo</span>
-                            </div>
-                          )}
-                          <div className="absolute inset-0 bg-black/40 opacity-0 group-hover:opacity-100 transition-opacity flex items-center justify-center">
-                            <span className="text-[10px] text-white font-medium">Edit</span>
-                          </div>
-                        </div>
-
-                        {/* Banner + Alt Banner */}
-                        <div className="h-24 flex-1 flex gap-1">
-                          {/* Banner */}
-                          <div
-                            onClick={() => {
-                              setActiveTab('images');
-                              setShowImageSearch({ type: 'banner', gameId: selectedGame.id });
-                              setActiveImageSearchTab('banner');
-                              setImageSearchQuery(selectedGame.title);
-                              handleSearchImages('banner');
-                            }}
-                            className="flex-1 relative group cursor-pointer border border-gray-700 rounded-lg overflow-hidden bg-gray-800 hover:border-green-500 transition-colors"
-                          >
-                            {(editedGame.bannerUrl || selectedGame.bannerUrl) ? (
-                              <img
-                                src={editedGame.bannerUrl || selectedGame.bannerUrl}
-                                alt="Banner"
-                                className="w-full h-full object-cover"
-                              />
-                            ) : (
-                              <div className="w-full h-full flex items-center justify-center">
-                                <span className="text-[8px] text-gray-600">Banner</span>
-                              </div>
-                            )}
-                            <div className="absolute inset-0 bg-black/40 opacity-0 group-hover:opacity-100 transition-opacity flex items-center justify-center">
-                              <span className="text-[10px] text-white font-medium">Edit</span>
-                            </div>
-                          </div>
-
-                          {/* Alt Banner */}
-                          <div
-                            onClick={() => {
-                              setActiveTab('images');
-                              setShowImageSearch({ type: 'alternativeBanner', gameId: selectedGame.id });
-                              setActiveImageSearchTab('alternativeBanner');
-                              setImageSearchQuery(selectedGame.title);
-                              handleSearchImages('alternativeBanner');
-                            }}
-                            className="flex-1 relative group cursor-pointer border border-gray-700 rounded-lg overflow-hidden bg-gray-800 hover:border-green-500 transition-colors"
-                          >
-                            {(editedGame.alternativeBannerUrl || selectedGame.alternativeBannerUrl) ? (
-                              <img
-                                src={editedGame.alternativeBannerUrl || selectedGame.alternativeBannerUrl}
-                                alt="Alt Banner"
-                                className="w-full h-full object-cover"
-                              />
-                            ) : (
-                              <div className="w-full h-full flex items-center justify-center">
-                                <span className="text-[8px] text-gray-600">Alt Banner</span>
-                              </div>
-                            )}
-                            <div className="absolute inset-0 bg-black/40 opacity-0 group-hover:opacity-100 transition-opacity flex items-center justify-center">
-                              <span className="text-[10px] text-white font-medium">Edit</span>
-                            </div>
-                          </div>
-                        </div>
-
-                        {/* Icon */}
-                        <div
-                          onClick={() => {
-                            setActiveTab('images');
-                            setShowImageSearch({ type: 'icon', gameId: selectedGame.id });
-                            setActiveImageSearchTab('icon');
-                            setImageSearchQuery(selectedGame.title);
-                            handleSearchImages('icon');
-                          }}
-                          className="h-24 w-24 relative group cursor-pointer border border-gray-700 rounded-lg overflow-hidden bg-gray-800 hover:border-green-500 transition-colors flex-shrink-0"
-                        >
-                          {(editedGame.iconUrl || selectedGame.iconUrl) ? (
-                            <div className="w-full h-full p-2 flex items-center justify-center">
-                              <img
-                                src={editedGame.iconUrl || selectedGame.iconUrl}
-                                alt="Icon"
-                                className="max-w-full max-h-full object-contain"
-                              />
-                            </div>
-                          ) : (
-                            <div className="w-full h-full flex items-center justify-center text-center p-1">
-                              <span className="text-[8px] text-gray-600">Icon</span>
-                            </div>
-                          )}
-                          <div className="absolute inset-0 bg-black/40 opacity-0 group-hover:opacity-100 transition-opacity flex items-center justify-center">
-                            <span className="text-[10px] text-white font-medium">Edit</span>
-                          </div>
-                        </div>
-                      </div>
-
-                      <div className="space-y-4">
-                        <div className="flex flex-col lg:flex-row gap-4">
-                          {/* Title Row with Fix Match */}
-                          <div className="flex-1">
-                            <label className="block text-xs font-medium text-gray-400 mb-1">Title</label>
-                            <div className="flex gap-2">
-                              <button
-                                type="button"
-                                onClick={() => {
-                                  const newLockedFields = { ...editedGame.lockedFields };
-                                  newLockedFields.title = !newLockedFields.title;
-                                  setEditedGame({ ...editedGame, lockedFields: newLockedFields });
-                                }}
-                                className={`flex items-center justify-center p-1.5 rounded transition-colors ${editedGame.lockedFields?.title ? 'bg-red-500/20 text-red-400 hover:bg-red-500/30' : 'bg-gray-700 text-gray-400 hover:bg-gray-600 hover:text-white'}`}
-                                title={editedGame.lockedFields?.title ? "Unlock Title" : "Lock Title"}
-                              >
-                                <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                                  {editedGame.lockedFields?.title ? (
-                                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 15v2m-6 4h12a2 2 0 002-2v-6a2 2 0 00-2-2H6a2 2 0 00-2 2v6a2 2 0 002 2zm10-10V7a4 4 0 00-8 0v4h8z" />
-                                  ) : (
-                                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M8 11V7a4 4 0 118 0m-4 8v2m-6 4h12a2 2 0 002-2v-6a2 2 0 00-2-2H6a2 2 0 00-2 2v6a2 2 0 002 2z" />
-                                  )}
-                                </svg>
-                              </button>
-                              <input
-                                type="text"
-                                value={showFixMatch ? metadataSearchQuery : editedGame.title}
-                                onChange={(e) => {
-                                  if (showFixMatch) {
-                                    setMetadataSearchQuery(e.target.value);
-                                  } else {
-                                    setEditedGame({ ...editedGame, title: e.target.value });
-                                  }
-                                }}
-                                onKeyDown={(e) => {
-                                  if (showFixMatch && e.key === 'Enter') {
-                                    handleFixMatchSearch();
-                                  }
-                                }}
-                                placeholder={showFixMatch ? "Enter game title to search..." : ""}
-                                className="flex-1 px-3 py-1.5 text-sm bg-gray-800 border border-gray-600 rounded text-white focus:outline-none focus:ring-1 focus:ring-blue-500"
-                                disabled={(showFixMatch && isSearchingMetadata) || editedGame.lockedFields?.title}
-                              />
-                              {showFixMatch && (
-                                <button
-                                  onClick={handleFixMatchSearch}
-                                  disabled={isSearchingMetadata}
-                                  className="px-4 py-1.5 text-sm bg-blue-600 hover:bg-blue-700 text-white rounded transition-colors disabled:opacity-50 flex items-center gap-2"
-                                >
-                                  {isSearchingMetadata ? (
-                                    <>
-                                      <svg className="animate-spin h-4 w-4 group- hover:animate-wobble group-hover:animate-wobble" fill="none" viewBox="0 0 24 24">
-                                        <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4" />
-                                        <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z" />
-                                      </svg>
-                                      Searching...
-                                    </>
-                                  ) : (
-                                    'Search'
-                                  )}
-                                </button>
-                              )}
-                              <button
-                                onClick={async () => {
-                                  const wasHidden = !showFixMatch;
-                                  setShowFixMatch(!showFixMatch);
-                                  if (wasHidden) {
-                                    // When opening, set search query to current title and auto-search
-                                    setMetadataSearchQuery(editedGame.title || selectedGame.title);
-                                    const query = editedGame.title || selectedGame.title;
-                                    if (query) {
-                                      setIsSearchingMetadata(true);
-                                      setMetadataSearchResults([]);
-                                      setError(null);
-                                      try {
-                                        // Note: searchGames returns an array directly, not a {success, results} wrapper
-                                        const response = await window.electronAPI.searchGames(query);
-                                        const results = Array.isArray(response) ? response : (response.results || []);
-
-                                        if (results.length === 0) {
-                                          setError('No matches found. Try a different search term.');
-                                          setMetadataSearchResults([]);
-                                        } else {
-                                          // Sort: Score > Date > Exact Match
-                                          const normalizedQuery = query.toLowerCase().trim();
-                                          const sortedResults = results.sort((a: any, b: any) => {
-                                            // 1. Score (assigned by backend)
-                                            const scoreA = a.score || 0;
-                                            const scoreB = b.score || 0;
-                                            if (scoreA !== scoreB) return scoreB - scoreA;
-
-                                            // 2. Release Date (Newest first)
-                                            const getDate = (r: any) => {
-                                              if (r.releaseDate) return typeof r.releaseDate === 'number' ? r.releaseDate * 1000 : new Date(r.releaseDate).getTime();
-                                              if (r.year) return new Date(r.year, 0, 1).getTime();
-                                              return 0;
-                                            };
-                                            const dateA = getDate(a);
-                                            const dateB = getDate(b);
-                                            if (dateA !== dateB && dateA > 0 && dateB > 0) return dateB - dateA;
-
-                                            // 3. Exact Match
-                                            const nameA = (a.title || a.name || "").toLowerCase().trim();
-                                            const nameB = (b.title || b.name || "").toLowerCase().trim();
-                                            if (nameA === normalizedQuery && nameB !== normalizedQuery) return -1;
-                                            if (nameA !== normalizedQuery && nameB === normalizedQuery) return 1;
-
-                                            return 0;
-                                          });
-                                          setMetadataSearchResults(sortedResults);
-                                        }
-                                      } catch (err) {
-                                        console.error('Error searching metadata:', err);
-                                        setError('Failed to search for games. Please try again.');
-                                      } finally {
-                                        setIsSearchingMetadata(false);
-                                      }
-                                    }
-                                  } else {
-                                    // When hiding, clear search results
-                                    setMetadataSearchResults([]);
-                                    setMetadataSearchQuery('');
-                                    setError(null);
-                                  }
-                                }}
-                                className="px-3 py-1.5 text-sm bg-purple-600 hover:bg-purple-700 text-white rounded transition-colors flex items-center gap-1.5"
-                              >
-                                <svg className="w-3.5 h-3.5 group- hover:animate-wobble group-hover:animate-wobble" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z" />
-                                </svg>
-                                {showFixMatch ? 'Hide' : 'Fix Match'}
-                              </button>
-                            </div>
-                          </div>
-
-
-                        </div>
-
-                        {/* Fix Match Results */}
-                        {showFixMatch && (
-                          <div className="space-y-2">
-                            {isSearchingMetadata && (
-                              <div className="flex items-center gap-2 text-sm text-gray-400">
-                                <svg className="animate-spin h-4 w-4 group- hover:animate-wobble group-hover:animate-wobble" fill="none" viewBox="0 0 24 24">
-                                  <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4" />
-                                  <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z" />
-                                </svg>
-                                Searching for metadata matches...
-                              </div>
-                            )}
-
-                            {metadataSearchResults.length > 0 && (
-                              <div className="max-h-96 overflow-y-auto">
-                                <div className="space-y-2">
-                                  {metadataSearchResults.map((result) => {
-                                    // Extract release date properly - show full date, not just year
-                                    let displayDate: string | undefined;
-                                    if (result.releaseDate) {
-                                      // Handle both Unix timestamp (seconds) and Date objects
-                                      if (typeof result.releaseDate === 'number') {
-                                        const date = new Date(result.releaseDate * 1000);
-                                        displayDate = date.toLocaleDateString('en-US', { year: 'numeric', month: 'short', day: 'numeric' });
-                                      } else if (typeof result.releaseDate === 'string') {
-                                        const date = new Date(result.releaseDate);
-                                        if (!isNaN(date.getTime())) {
-                                          displayDate = date.toLocaleDateString('en-US', { year: 'numeric', month: 'short', day: 'numeric' });
-                                        } else {
-                                          // Try parsing as ISO date string
-                                          displayDate = result.releaseDate;
-                                        }
-                                      }
-                                    } else if (result.year) {
-                                      // Fallback to year only if no full date available
-                                      displayDate = result.year.toString();
-                                    }
-
-                                    return (
-                                      <button
-                                        key={result.id}
-                                        onClick={() => handleSelectMetadataMatch({ id: result.id, source: result.source, steamAppId: result.steamAppId, title: result.title || result.name })}
-                                        disabled={isApplyingMetadata}
-                                        className="relative w-full text-left p-3 text-sm bg-gray-800 hover:bg-gray-700 rounded border border-gray-600 disabled:opacity-50 transition-colors flex items-center gap-3"
-                                      >
-                                        <div className="flex-1 min-w-0">
-                                          <p className="text-white font-medium text-sm truncate" title={result.title || result.name}>
-                                            {result.title || result.name}
-                                          </p>
-                                          <div className="flex items-center gap-2 mt-1">
-                                            <span className={`text-xs ${result.source === 'steam' ? 'text-blue-400' : 'text-gray-400'}`}>
-                                              {result.source === 'steam' ? 'Steam' : result.source === 'igdb' ? 'IGDB' : result.source === 'steamgriddb' ? 'SGDB' : result.source}
-                                            </span>
-                                            {result.steamAppId && (
-                                              <span className="text-xs text-gray-500">App ID: {result.steamAppId}</span>
-                                            )}
-                                            {displayDate && (
-                                              <span className="text-xs text-gray-400">ï¿½ï¿½ï¿½ {displayDate}</span>
-                                            )}
-                                          </div>
-                                        </div>
-                                        {isApplyingMetadata && (
-                                          <div className="absolute inset-0 bg-black/50 rounded flex items-center justify-center">
-                                            <svg className="animate-spin h-5 w-5 text-blue-500 group- hover:animate-wobble group-hover:animate-wobble" fill="none" viewBox="0 0 24 24">
-                                              <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4" />
-                                              <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z" />
-                                            </svg>
-                                          </div>
-                                        )}
-                                      </button>
-                                    );
-                                  })}
-                                </div>
-                              </div>
-                            )}
-                          </div>
-                        )}
-
-                        {/* Description and Categories Row */}
-                        <div className="flex flex-col lg:flex-row gap-4">
-                          {/* Description */}
-                          <div className="flex-1">
-                            <label className="block text-xs font-medium text-gray-400 mb-1">Description</label>
-                            <textarea
-                              value={editedGame.description || ''}
-                              onChange={(e) => setEditedGame({ ...editedGame, description: e.target.value })}
-                              className="w-full px-3 py-1.5 text-sm bg-gray-800 border border-gray-600 rounded text-white focus:outline-none focus:ring-1 focus:ring-blue-500"
-                              rows={4}
-                            />
-                          </div>
-
-                          {/* Categories - vertical scroll layout */}
-                          <div className="w-full lg:w-[35%] flex flex-col">
-                            <label className="block text-xs font-medium text-gray-400 mb-1">Categories</label>
-                            <div className="flex-1 p-2 bg-gray-800/50 rounded border border-gray-700 flex flex-col gap-2 max-h-[104px] overflow-y-auto scrollbar-thin scrollbar-thumb-gray-600 scrollbar-track-transparent">
-                              <div className="flex flex-wrap gap-1">
-                                {editedGame.categories?.map((category, index) => (
-                                  <span key={index} className="inline-flex items-center gap-1 px-2 py-0.5 text-[10px] font-medium bg-blue-900/30 text-blue-200 border border-blue-700/30 rounded-full group hover:border-blue-500/50 transition-colors">
-                                    {category}
-                                    <button
-                                      onClick={() => {
-                                        const newCategories = [...(editedGame.categories || [])];
-                                        newCategories.splice(index, 1);
-                                        setEditedGame({ ...editedGame, categories: newCategories });
-                                      }}
-                                      className="ml-0.5 text-blue-400 hover:text-white focus:outline-none rounded-full"
-                                    >
-                                      &times;
-                                    </button>
-                                  </span>
-                                ))}
-                              </div>
-                              <input
-                                type="text"
-                                value={newCategoryInput}
-                                onChange={(e) => setNewCategoryInput(e.target.value)}
-                                onKeyDown={(e) => {
-                                  if (e.key === 'Enter') {
-                                    e.preventDefault();
-                                    if (newCategoryInput.trim()) {
-                                      const currentCategories = editedGame.categories || [];
-                                      const newCat = newCategoryInput.trim();
-                                      if (!currentCategories.includes(newCat)) {
-                                        setEditedGame({
-                                          ...editedGame,
-                                          categories: [...currentCategories, newCat]
-                                        });
-                                        setNewCategoryInput('');
-                                      }
-                                    }
-                                  }
-                                }}
-                                placeholder="Add category..."
-                                className="w-full bg-transparent border-none text-xs text-white focus:outline-none placeholder-gray-500"
-                              />
-                            </div>
-                          </div>
-                        </div>
-
-                        {/* Metadata Grid - 5 Columns */}
-                        <div className="grid grid-cols-5 gap-2">
-                          {/* Platform */}
-                          {editedGame.platform && editedGame.platform !== 'other' && (
-                            <div>
-                              <label className="block text-[10px] uppercase tracking-wider font-semibold text-gray-500 mb-0.5">Platform</label>
-                              <input
-                                type="text"
-                                value={editedGame.platform}
-                                onChange={(e) => setEditedGame({ ...editedGame, platform: e.target.value })}
-                                className="w-full px-2 py-1 text-xs bg-gray-800 border border-gray-600 rounded text-white focus:outline-none focus:ring-1 focus:ring-blue-500"
-                              />
-                            </div>
-                          )}
-
-
-
-                          <div>
-                            <label className="block text-[10px] uppercase tracking-wider font-semibold text-gray-500 mb-0.5">Release Date</label>
-                            <input
-                              type="text"
-                              value={editedGame.releaseDate || ''}
-                              onChange={(e) => setEditedGame({ ...editedGame, releaseDate: e.target.value })}
-                              placeholder="YYYY-MM-DD"
-                              className="w-full px-2 py-1 text-xs bg-gray-800 border border-gray-600 rounded text-white focus:outline-none focus:ring-1 focus:ring-blue-500"
-                            />
-                          </div>
-                          <div>
-                            <label className="block text-[10px] uppercase tracking-wider font-semibold text-gray-500 mb-0.5">Age Rating</label>
-                            <input
-                              type="text"
-                              value={editedGame.ageRating || ''}
-                              onChange={(e) => setEditedGame({ ...editedGame, ageRating: e.target.value })}
-                              className="w-full px-2 py-1 text-xs bg-gray-800 border border-gray-600 rounded text-white focus:outline-none focus:ring-1 focus:ring-blue-500"
-                            />
-                          </div>
-                          <div>
-                            <label className="block text-[10px] uppercase tracking-wider font-semibold text-gray-500 mb-0.5">Series</label>
-                            <input
-                              type="text"
-                              value={editedGame.series || ''}
-                              onChange={(e) => setEditedGame({ ...editedGame, series: e.target.value })}
-                              className="w-full px-2 py-1 text-xs bg-gray-800 border border-gray-600 rounded text-white focus:outline-none focus:ring-1 focus:ring-blue-500"
-                            />
-                          </div>
-                          <div>
-                            <label className="block text-[10px] uppercase tracking-wider font-semibold text-gray-500 mb-0.5">Genres</label>
-                            <input
-                              type="text"
-                              value={editedGame.genres?.join(', ') || ''}
-                              onChange={(e) => setEditedGame({ ...editedGame, genres: e.target.value.split(',').map(g => g.trim()).filter(g => g) })}
-                              className="w-full px-2 py-1 text-xs bg-gray-800 border border-gray-600 rounded text-white focus:outline-none focus:ring-1 focus:ring-blue-500"
-                            />
-                          </div>
-                          <div>
-                            <label className="block text-[10px] uppercase tracking-wider font-semibold text-gray-500 mb-0.5">Developers</label>
-                            <input
-                              type="text"
-                              value={editedGame.developers?.join(', ') || ''}
-                              onChange={(e) => setEditedGame({ ...editedGame, developers: e.target.value.split(',').map(d => d.trim()).filter(d => d) })}
-                              className="w-full px-2 py-1 text-xs bg-gray-800 border border-gray-600 rounded text-white focus:outline-none focus:ring-1 focus:ring-blue-500"
-                            />
-                          </div>
-                          <div>
-                            <label className="block text-[10px] uppercase tracking-wider font-semibold text-gray-500 mb-0.5">Publishers</label>
-                            <input
-                              type="text"
-                              value={editedGame.publishers?.join(', ') || ''}
-                              onChange={(e) => setEditedGame({ ...editedGame, publishers: e.target.value.split(',').map(p => p.trim()).filter(p => p) })}
-                              className="w-full px-2 py-1 text-xs bg-gray-800 border border-gray-600 rounded text-white focus:outline-none focus:ring-1 focus:ring-blue-500"
-                            />
-                          </div>
-                          <div>
-                            <label className="block text-[10px] uppercase tracking-wider font-semibold text-gray-500 mb-0.5">Source</label>
-                            <input
-                              type="text"
-                              value={editedGame.source || ''}
-                              onChange={(e) => setEditedGame({ ...editedGame, source: e.target.value })}
-                              className="w-full px-2 py-1 text-xs bg-gray-800 border border-gray-600 rounded text-white focus:outline-none focus:ring-1 focus:ring-blue-500"
-                            />
-                          </div>
-                          <div className="col-span-2">
-                            <label className="block text-[10px] uppercase tracking-wider font-semibold text-gray-500 mb-0.5">Install Directory</label>
-                            <input
-                              type="text"
-                              value={editedGame.installationDirectory || ''}
-                              onChange={(e) => setEditedGame({ ...editedGame, installationDirectory: e.target.value })}
-                              className="w-full px-2 py-1 text-xs bg-gray-800 border border-gray-600 rounded text-white focus:outline-none focus:ring-1 focus:ring-blue-500"
-                            />
-                          </div>
-                          <div className="col-span-2">
-                            <label className="block text-[10px] uppercase tracking-wider font-semibold text-gray-500 mb-0.5">Executable Path</label>
-                            <input
-                              type="text"
-                              value={editedGame.exePath || ''}
-                              onChange={(e) => setEditedGame({ ...editedGame, exePath: e.target.value })}
-                              className="w-full px-2 py-1 text-xs bg-gray-800 border border-gray-600 rounded text-white focus:outline-none focus:ring-1 focus:ring-blue-500"
-                            />
-                          </div>
-                          <div>
-                            <label className="block text-[10px] uppercase tracking-wider font-semibold text-gray-500 mb-0.5">Install Size</label>
-                            <input
-                              type="text"
-                              value={editedGame.installSize ? `${Math.round(editedGame.installSize / 1024 / 1024 / 1024 * 100) / 100} GB` : ''}
-                              readOnly
-                              className="w-full px-2 py-1 text-xs bg-gray-800/50 border border-gray-600 rounded text-gray-400"
-                            />
-                          </div>
-
-                        </div>
-
-                        {/* Action Buttons */}
-                        <div className="flex gap-2 pt-2">
-                          <button
-                            onClick={handleSave}
-                            disabled={isSaving}
-                            className="flex-1 px-4 py-2 text-sm bg-blue-600 hover:bg-blue-700 text-white rounded transition-colors disabled:opacity-50"
-                          >
-                            {isSaving ? 'Saving...' : 'Save'}
-                          </button>
-                          <button
-                            onClick={() => {
-                              setExpandedGameId(null);
-                              setEditedGame(null);
-                              setShowFixMatch(false);
-                              setSelectedGameId(null);
-                              // Reset Fix Match state when cancelling edit
-                              setMetadataSearchResults([]);
-                              setMetadataSearchQuery('');
-                              setIsSearchingMetadata(false);
-                            }}
-                            className="px-4 py-2 text-sm bg-gray-600 hover:bg-gray-700 text-white rounded transition-colors"
-                          >
-                            Cancel
-                          </button>
-                          {onDeleteGame && (
-                            <button
-                              onClick={() => setShowDeleteConfirm(true)}
-                              disabled={isDeleting}
-                              className="px-4 py-2 text-sm bg-red-600 hover:bg-red-700 text-white rounded transition-colors disabled:opacity-50 flex items-center gap-2"
-                            >
-                              <svg className="w-4 h-4 group- hover:animate-wobble group-hover:animate-wobble" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />
-                              </svg>
-                              Delete
-                            </button>
-                          )}
-                        </div>
-                      </div>
-                    </div>
+                    <MetadataTab
+                      editedGame={editedGame}
+                      setEditedGame={setEditedGame}
+                      selectedGame={selectedGame}
+                      showFixMatch={showFixMatch}
+                      setShowFixMatch={setShowFixMatch}
+                      metadataSearchQuery={metadataSearchQuery}
+                      setMetadataSearchQuery={setMetadataSearchQuery}
+                      metadataSearchResults={metadataSearchResults}
+                      setMetadataSearchResults={setMetadataSearchResults}
+                      isSearchingMetadata={isSearchingMetadata}
+                      setIsSearchingMetadata={setIsSearchingMetadata}
+                      handleFixMatchSearch={handleFixMatchSearch}
+                      handleSelectMetadataMatch={handleSelectMetadataMatch}
+                      isApplyingMetadata={isApplyingMetadata}
+                      onSave={handleSave}
+                      isSaving={isSaving}
+                      onCancel={() => {
+                        setExpandedGameId(null);
+                        setEditedGame(null);
+                        setShowFixMatch(false);
+                        setSelectedGameId(null);
+                        setMetadataSearchResults([]);
+                        setMetadataSearchQuery('');
+                        setIsSearchingMetadata(false);
+                      }}
+                      onDeleteGame={onDeleteGame}
+                      setShowDeleteConfirm={setShowDeleteConfirm}
+                      isDeleting={isDeleting}
+                      setError={setError}
+                      setActiveTab={setActiveTab}
+                      setShowImageSearch={setShowImageSearch}
+                      setActiveImageSearchTab={setActiveImageSearchTab}
+                      setImageSearchQuery={setImageSearchQuery}
+                      handleSearchImages={handleSearchImages}
+                    />
                   )}
 
-                  {activeTab === 'links' && editedGame && (
-                    <div className="p-4 h-full overflow-y-auto flex flex-col">
-                      <div className="flex items-center justify-between mb-4">
-                        <label className="block text-xs font-semibold uppercase tracking-wider text-gray-400">Official Links</label>
-                        <div className="flex gap-2">
-                          <button
-                            type="button"
-                            onClick={async () => {
-                              if (!editedGame) return;
-                              setIsRefreshingLinks(true);
-                              setFoundLinks(null);
-                              setError(null);
-                              try {
-                                const result = await window.electronAPI.findLinks(editedGame.id);
-                                if (result.success) {
-                                  setFoundLinks(result.links);
-                                } else {
-                                  setError(result.error || 'Failed to find links');
-                                }
-                              } catch (error) {
-                                console.error('Failed to refresh links:', error);
-                                setError('An unexpected error occurred while searching for links');
-                              } finally {
-                                setIsRefreshingLinks(false);
-                              }
-                            }}
-                            disabled={isRefreshingLinks}
-                            className="text-xs text-green-400 hover:text-green-300 font-medium px-3 py-1.5 bg-green-500/10 hover:bg-green-500/20 rounded transition-colors disabled:opacity-50 flex items-center gap-2"
-                          >
-                            <svg className={`w-4 h-4 ${isRefreshingLinks ? 'animate-spin' : ''}`} fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15" />
-                            </svg>
-                            {isRefreshingLinks ? 'Searching...' : 'Refresh Links'}
-                          </button>
-                          <button
-                            type="button"
-                            onClick={() => {
-                              setEditedGame(prev => ({
-                                ...prev!,
-                                links: [...(prev!.links || []), { name: '', url: '' }]
-                              }));
-                            }}
-                            className="text-xs text-blue-400 hover:text-blue-300 font-medium px-3 py-1.5 bg-blue-500/10 hover:bg-blue-500/20 rounded transition-colors"
-                          >
-                            + Add Link
-                          </button>
-                        </div>
-                      </div>
-                      <div className="space-y-2 flex-1 min-h-0 overflow-y-auto pr-1 scrollbar-thin scrollbar-thumb-gray-600 scrollbar-track-transparent">
-                        {isRefreshingLinks ? (
-                          <div className="flex flex-col items-center justify-center py-8 gap-3">
-                            <div className="w-8 h-8 border-2 border-blue-500 border-t-transparent rounded-full animate-spin" />
-                            <div className="text-sm text-gray-400">Searching for Links...</div>
-                          </div>
-                        ) : foundLinks ? (
-                          <div className="bg-blue-500/5 border border-blue-500/20 rounded-lg p-3 space-y-3">
-                            <div className="flex items-center justify-between">
-                              <span className="text-xs font-semibold text-blue-400 uppercase tracking-wider">Found Links</span>
-                              <div className="flex gap-2">
-                                <button
-                                  type="button"
-                                  onClick={() => {
-                                    setEditedGame(prev => {
-                                      const existing = prev!.links || [];
-                                      const existingTypes = new Set(existing.map(l => (l.name || '').toLowerCase()));
-                                      const deduped = foundLinks.filter(l => !existingTypes.has((l.name || '').toLowerCase()));
-                                      return {
-                                        ...prev!,
-                                        links: [...existing, ...deduped]
-                                      };
-                                    });
-                                    setFoundLinks(null);
-                                  }}
-                                  className="text-xs bg-blue-600 hover:bg-blue-700 text-white px-2 py-1 rounded transition-colors"
-                                >
-                                  Apply All
-                                </button>
-                                <button
-                                  type="button"
-                                  onClick={() => setFoundLinks(null)}
-                                  className="text-xs bg-gray-700 hover:bg-gray-600 text-gray-300 px-2 py-1 rounded transition-colors"
-                                >
-                                  Cancel
-                                </button>
-                              </div>
-                            </div>
-                            <div className="space-y-1.5">
-                              {foundLinks.map((link, idx) => (
-                                <div key={idx} className="flex items-center justify-between text-xs bg-gray-800/50 px-2 py-1.5 rounded">
-                                  <span className="text-gray-300 font-medium truncate">{link.name}</span>
-                                  <span className="text-gray-500 truncate ml-2 max-w-[200px]">{link.url}</span>
-                                </div>
-                              ))}
-                              {foundLinks.length === 0 && <div className="text-xs text-gray-500 italic py-2">No new links found.</div>}
-                            </div>
-                          </div>
-                        ) : !editedGame.links || editedGame.links.length === 0 ? (
-                          <div className="text-sm text-gray-500 italic py-6">No links added. Use Refresh Links to fetch from IGDB or add manually.</div>
-                        ) : (
-                          <div className="space-y-2">
-                            {editedGame.links.map((link, idx) => {
-                              const iconKey = inferLinkKey(link.url, link.name);
-                              return (
-                                <div key={idx} className="flex gap-2 items-center">
-                                  <button
-                                    type="button"
-                                    onClick={() => setLinkIconPopupIndex(idx)}
-                                    title="Change icon"
-                                    className="w-9 h-9 flex-shrink-0 flex items-center justify-center rounded text-white border border-gray-600 hover:border-gray-500 transition-colors"
-                                  >
-                                    <LinkIcon iconKey={iconKey} className="w-[70%] h-[70%]" customIconUrl={link.iconUrl} />
-                                  </button>
-                                  <input
-                                    type="text"
-                                    value={link.name}
-                                    onChange={(e) => {
-                                      const newLinks = [...editedGame.links!];
-                                      newLinks[idx] = { ...newLinks[idx], name: e.target.value };
-                                      setEditedGame({ ...editedGame, links: newLinks });
-                                    }}
-                                    placeholder="Label (e.g. Steam)"
-                                    className="w-28 px-2 py-1.5 text-sm bg-gray-800 border border-gray-600 rounded text-white focus:outline-none focus:ring-1 focus:ring-blue-500"
-                                  />
-                                  <input
-                                    type="text"
-                                    value={link.url}
-                                    onChange={(e) => {
-                                      const newLinks = [...editedGame.links!];
-                                      newLinks[idx] = { ...newLinks[idx], url: e.target.value };
-                                      setEditedGame({ ...editedGame, links: newLinks });
-                                    }}
-                                    placeholder="URL"
-                                    className="flex-1 px-2 py-1.5 text-sm bg-gray-800 border border-gray-600 rounded text-white focus:outline-none focus:ring-1 focus:ring-blue-500"
-                                  />
-                                  <button
-                                    type="button"
-                                    onClick={() => {
-                                      const newLinks = [...editedGame.links!];
-                                      newLinks.splice(idx, 1);
-                                      setEditedGame({ ...editedGame, links: newLinks });
-                                    }}
-                                    className="px-3 py-1.5 text-xs text-red-400 hover:text-red-300 bg-red-500/10 hover:bg-red-500/20 rounded transition-colors shrink-0"
-                                  >
-                                    Remove
-                                  </button>
-                                </div>
-                              );
-                            })}
-                          </div>
-                        )}
-                      </div>
-                      <div className="flex gap-2 pt-4 mt-4 border-t border-gray-700 flex-shrink-0">
-                        <button
-                          onClick={handleSave}
-                          disabled={isSaving}
-                          className="flex-1 px-4 py-2 text-sm bg-blue-600 hover:bg-blue-700 text-white rounded transition-colors disabled:opacity-50"
-                        >
-                          {isSaving ? 'Saving...' : 'Save'}
-                        </button>
-                        <button
-                          onClick={() => {
-                            setExpandedGameId(null);
-                            setEditedGame(null);
-                            setShowFixMatch(false);
-                            setSelectedGameId(null);
-                            setMetadataSearchResults([]);
-                            setMetadataSearchQuery('');
-                            setIsSearchingMetadata(false);
-                          }}
-                          className="px-4 py-2 text-sm bg-gray-600 hover:bg-gray-700 text-white rounded transition-colors"
-                        >
-                          Cancel
-                        </button>
-                        {onDeleteGame && (
-                          <button
-                            onClick={() => setShowDeleteConfirm(true)}
-                            disabled={isDeleting}
-                            className="px-4 py-2 text-sm bg-red-600 hover:bg-red-700 text-white rounded transition-colors disabled:opacity-50 flex items-center gap-2"
-                          >
-                            <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />
-                            </svg>
-                            Delete
-                          </button>
-                        )}
-                      </div>
-                    </div>
+                  {activeTab === 'links' && (
+                    <LinksTab
+                      editedGame={editedGame}
+                      setEditedGame={setEditedGame}
+                      onSave={handleSave}
+                      onCancel={() => {
+                        setExpandedGameId(null);
+                        setEditedGame(null);
+                        setShowFixMatch(false);
+                        setSelectedGameId(null);
+                        setMetadataSearchResults([]);
+                        setMetadataSearchQuery('');
+                        setIsSearchingMetadata(false);
+                      }}
+                      isSaving={isSaving}
+                      onDeleteGame={onDeleteGame}
+                      setShowDeleteConfirm={setShowDeleteConfirm}
+                      isDeleting={isDeleting}
+                      setError={setError}
+                    />
                   )}
 
-                  {activeTab === 'modManager' && editedGame && (
-                    <div className="p-4">
-                      <div className="space-y-4">
-                        <div>
-                          <label className="block text-xs font-medium text-gray-400 mb-1">
-                            Mod Manager Link
-                          </label>
-                          <div className="flex gap-2">
-                            <input
-                              type="text"
-                              value={editedGame.modManagerUrl || ''}
-                              onChange={(e) => setEditedGame({ ...editedGame, modManagerUrl: e.target.value })}
-                              className="flex-1 px-3 py-1.5 text-sm bg-gray-800 border border-gray-600 rounded text-white placeholder-gray-500 focus:outline-none focus:ring-1 focus:ring-blue-500"
-                              placeholder="Enter mod manager URL or path (e.g., https://example.com/mod-manager)"
-                            />
-                            <button
-                              type="button"
-                              onClick={async () => {
-                                const path = await window.electronAPI.showOpenDialog();
-                                if (path) {
-                                  setEditedGame({ ...editedGame, modManagerUrl: path });
-                                }
-                              }}
-                              className="px-4 py-1.5 text-sm bg-gray-700 hover:bg-gray-600 text-white rounded transition-colors"
-                              title="Browse for mod manager executable"
-                            >
-                              Browse
-                            </button>
-                          </div>
-                          <p className="text-xs text-gray-500 mt-2">
-                            Enter the URL or path to your mod manager. This will appear in the game's context menu and bottom bar.
-                          </p>
-                        </div>
-
-                        {/* Action Buttons */}
-                        <div className="flex gap-2 pt-2">
-                          <button
-                            onClick={handleSave}
-                            disabled={isSaving}
-                            className="flex-1 px-4 py-2 text-sm bg-blue-600 hover:bg-blue-700 text-white rounded transition-colors disabled:opacity-50"
-                          >
-                            {isSaving ? 'Saving...' : 'Save'}
-                          </button>
-                          <button
-                            onClick={() => {
-                              setExpandedGameId(null);
-                              setEditedGame(null);
-                              setShowFixMatch(false);
-                              setSelectedGameId(null);
-                            }}
-                            className="px-4 py-2 text-sm bg-gray-600 hover:bg-gray-700 text-white rounded transition-colors"
-                          >
-                            Cancel
-                          </button>
-                        </div>
-                      </div>
-                    </div>
+                  {activeTab === 'modManager' && (
+                    <ModManagerTab
+                      editedGame={editedGame}
+                      setEditedGame={setEditedGame}
+                      onSave={handleSave}
+                      onCancel={() => {
+                        setExpandedGameId(null);
+                        setEditedGame(null);
+                        setShowFixMatch(false);
+                        setSelectedGameId(null);
+                      }}
+                      isSaving={isSaving}
+                    />
                   )}
                 </div>
               </>
@@ -3864,87 +2313,6 @@ export const GameManager: React.FC<GameManagerProps> = ({
         onRemove={handleRemoveMissingGames}
         onCancel={() => setShowRemoveDeletedDialog(false)}
       />
-
-      {/* Link icon change popup */}
-      {linkIconPopupIndex !== null && editedGame?.links?.[linkIconPopupIndex] && (
-        <div className="fixed inset-0 z-[9999] flex items-center justify-center bg-black/60" onClick={() => setLinkIconPopupIndex(null)}>
-          <div className="bg-gray-900 border border-gray-700 rounded-xl shadow-2xl p-4 w-[320px] max-w-[90vw]" onClick={e => e.stopPropagation()}>
-            <h3 className="text-sm font-semibold text-gray-200 mb-3">Change icon for &quot;{editedGame.links[linkIconPopupIndex].name || 'Link'}&quot;</h3>
-            <div className="flex flex-col gap-2">
-              <button
-                type="button"
-                onClick={() => {
-                  const query = getLinkIconSearchQuery(editedGame.links![linkIconPopupIndex].name);
-                  window.electronAPI.openExternal(`https://www.google.com/search?q=${query}`);
-                  setLinkIconPopupIndex(null);
-                }}
-                className="w-full px-3 py-2 text-sm bg-blue-600 hover:bg-blue-700 text-white rounded-lg transition-colors flex items-center justify-center gap-2"
-              >
-                <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z" />
-                </svg>
-                Search for icon in browser
-              </button>
-              <button
-                type="button"
-                onClick={() => linkIconFileInputRef.current?.click()}
-                className="w-full px-3 py-2 text-sm bg-gray-700 hover:bg-gray-600 text-white rounded-lg transition-colors flex items-center justify-center gap-2"
-              >
-                <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M4 16v1a3 3 0 003 3h10a3 3 0 003-3v-1m-4-8l-4-4m0 0L8 8m4-4v12" />
-                </svg>
-                Upload SVG icon
-              </button>
-              {editedGame.links[linkIconPopupIndex].iconUrl && (
-                <button
-                  type="button"
-                  onClick={() => {
-                    const newLinks = [...editedGame.links!];
-                    newLinks[linkIconPopupIndex] = { ...newLinks[linkIconPopupIndex], iconUrl: undefined };
-                    setEditedGame({ ...editedGame, links: newLinks });
-                    setLinkIconPopupIndex(null);
-                  }}
-                  className="w-full px-3 py-2 text-sm text-gray-300 hover:text-white bg-gray-800 hover:bg-gray-700 rounded-lg transition-colors"
-                >
-                  Remove custom icon
-                </button>
-              )}
-              <button
-                type="button"
-                onClick={() => setLinkIconPopupIndex(null)}
-                className="w-full px-3 py-2 text-sm text-gray-400 hover:text-gray-200 rounded-lg transition-colors"
-              >
-                Cancel
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
-      <input
-        ref={linkIconFileInputRef}
-        type="file"
-        accept=".svg"
-        className="hidden"
-        onChange={(e) => {
-          const file = e.target.files?.[0];
-          if (!file || linkIconPopupIndex === null || !editedGame?.links?.[linkIconPopupIndex]) {
-            e.target.value = '';
-            return;
-          }
-          const reader = new FileReader();
-          reader.onload = () => {
-            const dataUrl = reader.result as string;
-            const newLinks = [...editedGame.links!];
-            newLinks[linkIconPopupIndex] = { ...newLinks[linkIconPopupIndex], iconUrl: dataUrl };
-            setEditedGame({ ...editedGame, links: newLinks });
-            setLinkIconPopupIndex(null);
-          };
-          reader.readAsDataURL(file);
-          e.target.value = '';
-        }}
-      />
     </div >
   );
 };
-
-
